+++ noodles-sam/CHANGELOG.md
@@ -54,6 +54,9 @@
 
   * sam/header: Change `Comments` to a list of byte strings (`Vec<Vec<u8>>`).
 
+  * sam/header/record/value/map: Change `OtherFields` values to byte strings
+    (`Vec<u8>`).
+
   * sam/io/reader: Rename record to record buf and lazy record to record.
 
     This changes the following: